Keep an history of messages already displayed.
If the display radius of an user changes, we don't want previous
messages to disapear, which is confusing to the user. We so store
displayed messages uuids in the session and add them to the message
database request. In case the position is set and was previously
retrieved from Geoip we clear the history as Geoip can be very far from
the user position.